moved Loading component into a different recipe outside withSpinner
