+++ cookbook-react/src/recipes/spinners/SpinkitSpinner/constants.ts
