+++ cookbook-react/src/recipes/spinners/SpinkitSpinner/cookbook.json
@@ -1,0 +1,10 @@
+{
+  "thumbnail": {
+    "type": "iframe",
+    "url": "https://cookbook-react.now.sh/?category=spinners&component=SpinkitSpinner-1&compact=true"
+  },
+  "detail": {
+    "type": "iframe",
+    "url": "https://cookbook-react.now.sh/?category=spinners&component=SpinkitSpinner-1&compact=true"
+  }
+}
