+++ cookbook-react/src/recipes/spinners/SpinkitSpinner/index.test.tsx
